talentdashboard/requirements/base.txt
Django==1.5.4
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
djangorestframework==2.3.8
gunicorn==18.0
psycopg2==2.5.1
python-dateutil==2.1
python-mimeparse==0.1.4
six==1.4.1
static==0.4
wsgiref==0.1.2
pil==1.1.6
